requirements.txt
openai>=1.40.0
requests>=2.32.0
PyYAML>=6.0.1
matplotlib>=3.8.0
pytest>=8.0.0
